feat: complete CI/CD setup and code quality improvements
- Add comprehensive GitHub Actions CI workflow for Rust eBPF project
- Configure build pipeline with eBPF toolchain (nightly, bpf-linker, LLVM)
- Implement efficient caching strategy for dependencies and build artifacts
- Enable release-mode builds and comprehensive testing
- Fix all clippy warnings and errors:
  * Remove unnecessary cast in cookie generation
  * Rename ConnectionId::new to generate (fix new_ret_no_self warning)
  * Replace redundant closures with function references
  * Add dead_code allows for test helper functions
- Add crates.io publishing support with proper package metadata
- Streamline CI workflow for efficiency:
  * Combine cache steps into single efficient cache
  * Remove redundant operations from publish job
  * Optimize build steps and reduce duplication
- Configure publish job to run on main branch pushes
- Switch to git dependencies approach (remove crates.io publishing)
- Maintain code quality with formatting and linting checks

This establishes a robust CI/CD pipeline for the QUIC eBPF router library
with automated testing, quality checks, and efficient dependency management.

+++ common/src/lib.rs
@@ -13,14 +13,14 @@ pub struct Cookie;
 
 impl Cookie {
     /// Generate a cookie from generation and worker index
-    /// 
+    ///
     /// # Arguments
     /// * `generation` - Generation counter (0-31, only lower 5 bits used)
     /// * `worker_idx` - Worker/socket index (0-255)
-    /// 
+    ///
     /// # Returns
     /// A 16-bit cookie value with embedded checksum
-    /// 
+    ///
     /// # Example
     /// ```
     /// let cookie = common::Cookie::generate(0, 42);
@@ -30,16 +30,16 @@ impl Cookie {
     pub const fn generate(generation: u8, worker_idx: u8) -> u16 {
         let g = (generation & 0x1F) as u16; // 5 bits
         let idx = worker_idx as u16; // 8 bits
-        let checksum = ((g + idx) & 0x7) as u16; // 3 bits
-        
+        let checksum = (g + idx) & 0x7; // 3 bits
+
         (g << 11) | (idx << 3) | checksum
     }
-    
+
     /// Validate a cookie's checksum
-    /// 
+    ///
     /// # Arguments
     /// * `cookie` - The 16-bit cookie value to validate
-    /// 
+    ///
     /// # Returns
     /// `true` if the checksum is valid, `false` otherwise
     #[inline]
@@ -48,16 +48,16 @@ impl Cookie {
         let idx = (cookie >> 3) & 0xff;
         let chksum = cookie & 0x7;
         let sum = generation + idx;
-        
+
         chksum == (sum & 0x7)
     }
-    
+
     /// Extract the generation from a cookie
     #[inline]
     pub const fn get_generation(cookie: u16) -> u8 {
         (cookie >> 11) as u8
     }
-    
+
     /// Extract the worker index from a cookie
     #[inline]
     pub const fn get_worker_idx(cookie: u16) -> u8 {
